refactor: convert pagination as component
